requirements.txt
accelerate==0.24.1
aioboto3==12.0.0
aiobotocore==2.7.0
aiofiles==23.2.1
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
aspose-words==23.12.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.2.2
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.2.post2
blinker==1.7.0
boto3==1.28.64
botocore==1.31.64
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
clickhouse-connect==0.6.23
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.14.6
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
docx2txt==0.8
duckdb==0.9.2
effdet==0.4.1
emoji==2.8.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
faiss-cpu==1.7.4
Faker==19.13.0
fastapi==0.106.0
favicon==0.7.0
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.24.0
googleapis-common-protos==1.61.0
gradio_client==0.7.0
grpcio==1.59.3
h11==0.14.0
hnswlib==0.8.0
htbuilder==0.6.2
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
humbug==0.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
InstructorEmbedding==1.0.1
iopath==0.1.10
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.12
langdetect==1.0.9
langsmith==0.0.82
layoutparser==0.3.4
libdeeplake==0.0.90
llama_cpp_python==0.1.83
lxml==4.9.3
lz4==4.3.2
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.1
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openapi-schema-pydantic==1.2.4
opencv-python==4.8.1.78
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pathos==0.3.1
pdf-to-markdown==0.1.0
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
Pillow==10.1.0
pluggy==1.3.0
portalocker==2.8.2
posthog==3.0.2
pox==0.3.3
ppft==1.7.6.7
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.3.1
pypandoc==1.12
pyparsing==3.1.1
pypdfium2==4.23.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pytesseract==0.3.10
pytest==7.4.3
pytest-subtests==0.11.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3.post1
PyYAML==6.0.1
rank-bm25==0.2.2
rapidfuzz==3.5.2
referencing==0.30.2
regex==2023.10.3
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rouge==1.0.1
rpds-py==0.12.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sseclient-py==1.8.0
st-annotated-text==4.0.1
stack-data==0.6.3
starlette==0.27.0
streamlit==1.28.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.10
together==0.2.7
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.1
transformers==4.35.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.10.29
unstructured-inference==0.7.11
unstructured.pytesseract==0.3.12
urllib3==1.26.6
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
watermark==2.4.3
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
zstandard==0.22.0
cohere
flask-mail